# tlias-web-management/assert/Web开发-AI.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Web开发-AI",
    "description": "新版本的Web开发的接口文档",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/clazzs/{id}": {
      "get": {
        "summary": "根据ID查询",
        "x-apifox-folder": "班级管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID查询班级信息",
        "operationId": "getInfoUsingGET",
        "tags": [
          "班级管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {
                    "01H7YTX04J7GX93MGMS37EFYGC": {
                      "$ref": "#/components/schemas/Result"
                    }
                  },
                  "x-apifox-orders": [
                    "01H7YTX04J7GX93MGMS37EFYGC"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32"
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "nullable": true
                    },
                    "data": {
                      "title": "响应数据",
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/Clazz"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "x-apifox-ignore-properties": [
                    "code",
                    "msg",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410787-run"
      },
      "delete": {
        "summary": "删除班级",
        "x-apifox-folder": "班级管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID删除班级信息",
        "operationId": "deleteUsingDELETE",
        "tags": [
          "班级管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410788-run"
      }
    },
    "/depts": {
      "get": {
        "summary": "查询所有部门",
        "x-apifox-folder": "部门管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "查询所有部门",
        "operationId": "listUsingGET",
        "tags": [
          "部门管理"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJuYW1lIjoi5a6L5rGfIiwiaWQiOjIsInVzZXJuYW1lIjoic29uZ2ppYW5nIiwiZXhwIjoxNzA0NzU5MjE3fQ.isGb5c2pqqhH7PVC1Cm0AUaxskObBExE0SsJtmZs-7c",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {
                    "01H7YPXV7VMRGR32FMEW9ED0JH": {
                      "$ref": "#/components/schemas/Result"
                    }
                  },
                  "x-apifox-orders": [
                    "01H7YPXV7VMRGR32FMEW9ED0JH"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32"
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "nullable": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "x-apifox-refs": {
                          "01H80F03TV2N4W2ZZ29J2K33D7": {
                            "$ref": "#/components/schemas/Dept",
                            "x-apifox-overrides": {
                              "id": {
                                "type": "integer",
                                "title": "部门ID",
                                "format": "int32",
                                "nullable": true
                              },
                              "name": {
                                "type": "string",
                                "title": "部门名称",
                                "nullable": true
                              },
                              "createTime": {
                                "type": "string",
                                "title": "创建时间",
                                "format": "date-time",
                                "nullable": true
                              },
                              "updateTime": {
                                "type": "string",
                                "title": "修改时间",
                                "format": "date-time",
                                "nullable": true
                              }
                            }
                          }
                        },
                        "x-apifox-orders": [
                          "01H80F03TV2N4W2ZZ29J2K33D7"
                        ],
                        "properties": {
                          "id": {
                            "type": "integer",
                            "title": "部门ID",
                            "format": "int32",
                            "nullable": true
                          },
                          "name": {
                            "type": "string",
                            "title": "部门名称",
                            "nullable": true
                          },
                          "createTime": {
                            "type": "string",
                            "title": "创建时间",
                            "format": "date-time",
                            "nullable": true
                          },
                          "updateTime": {
                            "type": "string",
                            "title": "修改时间",
                            "format": "date-time",
                            "nullable": true
                          }
                        },
                        "x-apifox-ignore-properties": [
                          "id",
                          "name",
                          "createTime",
                          "updateTime"
                        ]
                      },
                      "title": "响应数据"
                    }
                  },
                  "x-apifox-ignore-properties": [
                    "code",
                    "msg",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410789-run"
      },
      "post": {
        "summary": "添加部门",
        "x-apifox-folder": "部门管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "添加部门数据",
        "operationId": "saveUsingPOST_1",
        "tags": [
          "部门管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "title": "部门名称"
                  }
                },
                "x-apifox-refs": {},
                "x-apifox-orders": [
                  "name"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": {
                "name": "近派使算他"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410790-run"
      },
      "put": {
        "summary": "修改部门",
        "x-apifox-folder": "部门管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "修改部门",
        "operationId": "updateUsingPUT_1",
        "tags": [
          "部门管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "title": "部门ID",
                    "format": "int32",
                    "nullable": true
                  },
                  "name": {
                    "type": "string",
                    "title": "部门名称",
                    "nullable": true
                  }
                },
                "x-apifox-refs": {},
                "x-apifox-orders": [
                  "id",
                  "name"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": {
                "name": "后勤部",
                "id": 10
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410791-run"
      },
      "delete": {
        "summary": "删除部门",
        "x-apifox-folder": "部门管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID删除部门数据",
        "operationId": "deleteUsingDELETE_1",
        "tags": [
          "部门管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "部门ID",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410793-run"
      }
    },
    "/depts/{id}": {
      "get": {
        "summary": "根据ID查询",
        "x-apifox-folder": "部门管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID查询部门",
        "operationId": "getByIdUsingGET",
        "tags": [
          "部门管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "10",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {
                    "01H7YQ828NM86WJCYT9AQGVXRH": {
                      "$ref": "#/components/schemas/Result"
                    }
                  },
                  "x-apifox-orders": [
                    "01H7YQ828NM86WJCYT9AQGVXRH"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32"
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "nullable": true
                    },
                    "data": {
                      "title": "响应数据",
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/Dept",
                          "x-apifox-overrides": {
                            "id": {
                              "type": "integer",
                              "title": "部门ID",
                              "format": "int32",
                              "nullable": true
                            },
                            "name": {
                              "type": "string",
                              "title": "部门名称",
                              "nullable": true
                            },
                            "createTime": {
                              "type": "string",
                              "title": "创建时间",
                              "format": "date-time",
                              "nullable": true
                            },
                            "updateTime": {
                              "type": "string",
                              "title": "修改时间",
                              "format": "date-time",
                              "nullable": true
                            }
                          }
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "x-apifox-ignore-properties": [
                    "code",
                    "msg",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410792-run"
      }
    },
...
    "/students/{id}": {
      "get": {
        "summary": "根据ID查询",
        "x-apifox-folder": "学生管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID查询",
        "operationId": "getInfoUsingGET_2",
        "tags": [
          "学生管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {
                    "01H7Z1NS5AJB5W2JG6FFQ3NGYE": {
                      "$ref": "#/components/schemas/Result"
                    }
                  },
                  "x-apifox-orders": [
                    "01H7Z1NS5AJB5W2JG6FFQ3NGYE"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32"
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "nullable": true
                    },
                    "data": {
                      "title": "响应数据",
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/Student"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "x-apifox-ignore-properties": [
                    "code",
                    "msg",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410809-run"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Clazz": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "title": "ID",
            "description": "ID"
          },
          "name": {
            "type": "string",
            "title": "班级名称",
            "description": "班级"
          },
          "room": {
            "type": "string",
            "title": "班级教室",
            "description": "班级教室"
          },
          "beginDate": {
            "type": "string",
            "format": "date",
            "title": "开课时间",
            "description": "开课时间"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "title": "结课时间",
            "description": "结课时间"
          },
          "masterId": {
            "type": "integer",
            "format": "int32",
            "title": "班主任id",
            "description": "班主任id, 关联员工表id"
          },
          "subject": {
            "type": "integer",
            "format": "int32",
            "title": "学科",
            "description": "1:java, 2:前端, 3:大数据, 4:Python, 5:Go, 6:嵌入式"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "title": "创建时间",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "title": "修改时间",
            "description": "修改时间"
          }
        },
        "title": "Clazz",
        "x-apifox-orders": [
          "id",
          "name",
          "room",
          "beginDate",
          "endDate",
          "masterId",
          "subject",
          "createTime",
          "updateTime"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
...
      "Dept": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "title": "部门ID",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "title": "部门名称",
            "nullable": true
          },
          "createTime": {
            "type": "string",
            "title": "创建时间",
            "format": "date-time",
            "nullable": true
          },
          "updateTime": {
            "type": "string",
            "title": "修改时间",
            "format": "date-time",
            "nullable": true
          }
        },
        "title": "Dept",
        "x-apifox-orders": [
          "id",
          "name",
          "createTime",
          "updateTime"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "title": "状态码",
            "description": "1: 成功 , 0: 失败",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "title": "错误提示信息",
            "nullable": true
          },
          "data": {
            "type": "object",
            "properties": {},
            "title": "响应数据",
            "x-apifox-orders": [],
            "x-apifox-ignore-properties": [],
            "nullable": true
          }
        },
        "title": "Result",
        "x-apifox-orders": [
          "code",
          "msg",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "开发环境"
    }
  ]
}